requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
Pillow==10.0.0
plyer==2.1.0
pystray==0.19.4
pywin32==306
requests==2.31.0
schedule==1.2.0
six==1.16.0
toml==0.10.2
urllib3==2.0.4
